requirements.txt
APScheduler==3.11.0
channels==4.2.2
cryptography>=41.0.7,<46.0.0
Django==5.2
django_apscheduler==0.7.0
django_minio_backend==3.8.0
django_minio_storage==0.5.8
djangorestframework==3.16.0
django_filter==25.1
drf_extensions==0.8.0
minio==7.2.15
Pillow>=10.0.0,<12.0.0
PyJWT==2.10.1
urllib3>=2.0.0,<3.0.0
daphne==4.2.0
mysqlclient>=2.2.0,<3.0.0
pytest==8.3.4
pytest-django==4.9.0